pyproject.toml
[project]
name = "hp-cccb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=0.36.0",
    "langchain[huggingface]>=1.2.0",
    "langchain-huggingface>=1.2.0",
    "langgraph>=1.0.5",
    "langsmith>=0.5.1",
    "python-dotenv>=1.2.1",
    "transformers>=4.57.3",
    "sentence-transformers>=5.2.0",
    "ipywidgets>=8.1.8",
    "langchain-core>=1.2.5",
    "langchain-community>=0.4.1",
    "pypdf>=6.5.0",
    "pymupdf>=1.26.7",
    "unstructured[pdf]>=0.18.21",
    "langchain-opendataloader-pdf>=1.0.1",
    "langchain-ollama>=1.0.1",
    "fastapi[standard]>=0.128.0",
    "faiss-cpu>=1.13.2",
]
